services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiodataloader>=0.4.3",
    "aniso8601>=10.0.1",
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.8",
    "fastapi>=0.133.1",
    "gunicorn>=25.1.0",
    "pyotp>=2.9.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "shared",
    "strawberry-graphql[fastapi]>=0.307.1",
    "uvicorn>=0.41.0",
]

[tool.uv.sources]
shared = { workspace = true, editable = true }
